+++ col.go
@@ -4,8 +4,8 @@ import (
 	"image"
 	"sort"
 
-	"github.com/rjkroege/edwood/frame"
 	"github.com/rjkroege/edwood/internal/draw"
+	"github.com/rjkroege/edwood/internal/frame"
 )
 
 var (
