+++ internal/frame/box_test.go
