+++ internal/frame/boxtest_helpers_test.go
