+++ internal/frame/delete.go
