+++ internal/frame/str.go
