+++ internal/frame/util.go
