+++ scrl.go
@@ -5,8 +5,8 @@ import (
 	"image"
 	"time"
 
-	"github.com/rjkroege/edwood/frame"
 	"github.com/rjkroege/edwood/internal/draw"
+	"github.com/rjkroege/edwood/internal/frame"
 )
 
 var scrtmp *draw.Image
